fixed PR code style suggestions.
