--- src/constants/index.ts
@@ -1,2 +1,0 @@
-export const SING_UP_URL = 'users';
-export const SING_IN_URL = 'sign_in';
