+++ src/services/UserService.ts
@@ -1,5 +1,6 @@
-import { SignupSuccess, SignupError, User } from '../typings/user';
-import { SING_UP_URL } from '../constants';
+import { SignupSuccess, SignupError, User } from '../utils/types';
 import api from '../config/api';
+
+const SING_UP_URL = 'users';
 
 export const signUp = (payload: User) => api.post<SignupSuccess, SignupError>(SING_UP_URL, payload);
