+++ src/utils/types.ts
@@ -1,1 +1,25 @@
 export type Nullable<T> = T | null;
+
+export interface User {
+  id?: number;
+  firstName: string;
+  lastName: string;
+  email: string;
+  password: string;
+  locale?: string;
+}
+
+export interface SignupSuccess {
+  id: number;
+  firstName: string;
+  lastName: string;
+  email: string;
+  locale: string;
+}
+
+export interface SignupError {
+  status: string;
+  errors: {
+    fullMessages: string[];
+  };
+}
